Add singleplayer mode (#23)

+++ src/entities/game.rs
@@ -8,7 +8,7 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
-    pub chat_id: i32,
+    pub chat_id: i64,
     pub game_mode: Gamemodes,
     pub active: bool,
 }
